Refactor to place Context2D and Matrix in separate modules
