+++ src/dummy/toga_dummy/factory.py
@@ -5,11 +5,13 @@ from .window import Window
 from .widgets.box import Box
 from .widgets.button import Button
 from .command import Command
-from .widgets.canvas import Canvas, Context2D, Matrix
+from .widgets.canvas import Canvas
+from .widgets.context2d import Context2D
 from .widgets.icon import Icon
 from .widgets.image import *
 from .widgets.imageview import *
 from .widgets.label import Label
+from .widgets.matrix import Matrix
 from .widgets.multilinetextinput import *
 from .widgets.navigationview import *
 from .widgets.numberinput import *
