+++ src/gtk/toga_gtk/widgets/matrix.py
@@ -1,0 +1,11 @@
+import cairo
+from .base import Widget
+
+
+class Matrix(Widget):
+    def create(self):
+        self.native = cairo.Matrix()
+        self.native.interface = self.interface
+
+    def transform_point(self, x, y):
+        return self.native.transform_point(x, y)
